updated delete to check chemical association
SVN-Revision: 12116

+++ gov/nih/nci/cananolab/ui/particle/FunctionalizingEntityAction.java
@@ -8,7 +8,7 @@ package gov.nih.nci.cananolab.ui.particle;
  * @author pansu
  */
 
-/* CVS $Id: FunctionalizingEntityAction.java,v 1.24 2008-05-04 09:47:45 pansu Exp $ */
+/* CVS $Id: FunctionalizingEntityAction.java,v 1.25 2008-05-05 14:20:07 pansu Exp $ */
 
 import gov.nih.nci.cananolab.dto.common.UserBean;
 import gov.nih.nci.cananolab.dto.particle.ParticleBean;
@@ -215,16 +215,26 @@ public class FunctionalizingEntityAction extends BaseAnnotationAction {
 				.getDisplayNameToClassNameLookup(
 						request.getSession().getServletContext()), user
 				.getLoginName());
-		compositionService.deleteFunctionalizingEntity(entityBean
-				.getDomainEntity());
+		boolean canDelete = compositionService
+				.checkChemicalAssociationBeforeDelete(entityBean);
 		ActionMessages msgs = new ActionMessages();
-		ActionMessage msg = new ActionMessage(
-				"message.deleteFunctionalizingEntity");
-		msgs.add(ActionMessages.GLOBAL_MESSAGE, msg);
-		saveMessages(request, msgs);
-		ActionForward forward = mapping.findForward("success");
-		setupDataTree(theForm, request);
-		return forward;
+		if (canDelete) {
+			compositionService.deleteFunctionalizingEntity(entityBean
+					.getDomainEntity());
+			ActionMessage msg = new ActionMessage(
+					"message.deleteFunctionalizingEntity");
+			msgs.add(ActionMessages.GLOBAL_MESSAGE, msg);
+			saveMessages(request, msgs);
+			setupDataTree(theForm, request);
+			return mapping.findForward("success");
+		} else {
+			ActionMessage msg = new ActionMessage(
+					"error.deleteFunctionalizingEntityWithChemicalAssociation");
+			msgs.add(ActionMessages.GLOBAL_MESSAGE, msg);
+			saveErrors(request, msgs);
+			setupDataTree(theForm, request);
+			return mapping.getInputForward();
+		}
 	}
 
 	protected FunctionalizingEntityBean[] prepareCopy(
